+++ lib/agent_daemon/daemon.rb
@@ -88,6 +88,8 @@ module AgentDaemon
         -> { Runner::Pachca.new(runner_config, message_dir, project_path, @shutdown_flag) }
       when "github"
         -> { Runner::GitHub.new(runner_config, message_dir, project_path, @shutdown_flag) }
+      when "basecamp"
+        -> { Runner::Basecamp.new(runner_config, message_dir, project_path, @shutdown_flag) }
       else
         raise ArgumentError, "Unknown trigger type #{type.inspect} in runner #{runner_config['name'].inspect}"
       end